Make workspace cleanup optional

+++ src/main/java/org/jboss/qa/jenkins/test/executor/phase/cleanup/CleanUpPhaseProcessor.java
@@ -18,12 +18,13 @@ public class CleanUpPhaseProcessor extends PhaseDefinitionProcessor {
 
 	public void execute() {
 		log.debug("@{} - {}", CleanUp.class.getName(), cleanUp.id());
-		try {
-			// TODO(vchalupa): Make a decision about this, some parameter in annotation may be set...
-			// TODO(vchalupa): Without clean the unpacked server home can not be recognized
-			FileUtils.deleteDirectory(JenkinsTestExecutor.WORKSPACE);
-		} catch (IOException e) {
-			e.printStackTrace();
+
+		if (cleanUp.cleanWorkspace()) {
+			try {
+				FileUtils.deleteDirectory(JenkinsTestExecutor.WORKSPACE);
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
 		}
 	}
 }
